Move permission entity-id types into metabase-types
DatabaseEntityId, SchemaEntityId, TableEntityId, EntityId,
EntityWithGroupId, PermissionSubject, and SpecialGroupType are pure data
shapes used by the plugin system (PLUGIN_ADMIN_PERMISSIONS_* signatures),
the admin permissions editor, and enterprise permission features.

Move them to metabase-types/api/permissions.ts (where DataPermission
already lives) and update all admin/enterprise consumers to import them
from metabase-types directly.

The permissions EntityId is renamed to PermissionEntityId, both to avoid
colliding with the comments EntityId in the metabase-types/api barrel and
to disambiguate the two types; all usage sites are renamed accordingly.

Fixes 1 module boundary violation (shared/plugins -> feature/admin).
