+++ enterprise/frontend/src/metabase-enterprise/advanced_permissions/reducer.ts
@@ -7,7 +7,6 @@ import {
   SAVE_DATA_PERMISSIONS,
   UPDATE_DATA_PERMISSION,
 } from "metabase/admin/permissions/permissions";
-import type { EntityId } from "metabase/admin/permissions/types";
 import {
   DATABASES_BASE_PATH,
   GROUPS_BASE_PATH,
@@ -17,10 +16,11 @@ import {
   DataPermissionValue,
   type GroupId,
   type Impersonation,
+  type PermissionEntityId,
 } from "metabase-types/api";
 
 export const getImpersonatedPostAction = (
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   groupId: GroupId,
   view: "database" | "group",
 ) =>
