+++ enterprise/frontend/src/metabase-enterprise/advanced_permissions/utils.ts
@@ -1,9 +1,8 @@
-import type { EntityId } from "metabase/admin/permissions/types";
 import {
   getDatabaseFocusPermissionsUrl,
   getGroupFocusPermissionsUrl,
 } from "metabase/admin/permissions/utils/urls";
-import type { GroupId } from "metabase-types/api";
+import type { GroupId, PermissionEntityId } from "metabase-types/api";
 
 import type { ImpersonationModalParams } from "./types";
 
@@ -22,7 +21,7 @@ export const getImpersonatedDatabaseId = ({
 };
 
 const getDatabaseViewImpersonationModalUrl = (
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   groupId: GroupId,
 ) => {
   const baseUrl = getDatabaseFocusPermissionsUrl(entityId);
@@ -30,7 +29,7 @@ const getDatabaseViewImpersonationModalUrl = (
 };
 
 const getGroupViewImpersonationModalUrl = (
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   groupId: GroupId,
 ) => {
   const baseUrl = getGroupFocusPermissionsUrl(groupId);
@@ -39,7 +38,7 @@ const getGroupViewImpersonationModalUrl = (
 };
 
 export const getEditImpersonationUrl = (
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   groupId: GroupId,
   view: "database" | "group",
 ) =>
