+++ enterprise/frontend/src/metabase-enterprise/feature_level_permissions/permission-management/data-model-permission.unit.spec.ts
@@ -1,8 +1,8 @@
-import type { SpecialGroupType } from "metabase/admin/permissions/types";
 import {
   DataPermissionValue,
   type Group,
   type GroupsPermissions,
+  type SpecialGroupType,
 } from "metabase-types/api";
 
 import {
