+++ enterprise/frontend/src/metabase-enterprise/feature_level_permissions/permission-management/details-permission.ts
@@ -8,16 +8,16 @@ import {
 } from "metabase/admin/permissions/selectors/confirmations";
 import {
   DataPermissionType,
-  type EntityId,
   type PermissionOption,
   type PermissionSectionConfig,
-  type PermissionSubject,
 } from "metabase/admin/permissions/types";
 import {
   DataPermission,
   DataPermissionValue,
   type Group,
   type GroupsPermissions,
+  type PermissionEntityId,
+  type PermissionSubject,
 } from "metabase-types/api";
 
 export const DETAILS_PERMISSION_OPTIONS: Record<string, PermissionOption> = {
@@ -51,7 +51,7 @@ const getDetailsPermission = (
   DETAILS_PERMISSION_OPTIONS.no.value;
 
 export const buildDetailsPermission = (
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   groupId: number,
   isAdmin: boolean,
   isExternal: boolean,
