+++ frontend/src/metabase/admin/permissions/permissions.ts
@@ -40,6 +40,7 @@ import type {
   CollectionPermissionsGraph,
   GroupId,
   GroupsPermissions,
+  PermissionEntityId,
   PermissionsGraph,
 } from "metabase-types/api";
 
@@ -48,7 +49,6 @@ import {
   DataPermission,
   DataPermissionType,
   type DataPermissionValue,
-  type EntityId,
   type PermissionSectionConfig,
 } from "./types";
 import {
@@ -191,7 +191,7 @@ export interface UpdateDataPermissionParams {
     "type" | "permission" | "postActions"
   >;
   value: DataPermissionValue;
-  entityId: EntityId;
+  entityId: PermissionEntityId;
   view: "database" | "group";
 }
 interface UpdateDataPermissionPayload {
@@ -202,7 +202,7 @@ interface UpdateDataPermissionPayload {
   >;
   value: DataPermissionValue;
   metadata: Metadata;
-  entityId: EntityId;
+  entityId: PermissionEntityId;
 }
 export const UPDATE_DATA_PERMISSION =
   "metabase/admin/permissions/UPDATE_DATA_PERMISSION";
