+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/schemas.ts
@@ -9,7 +9,12 @@ import {
   PLUGIN_FEATURE_LEVEL_PERMISSIONS,
 } from "metabase/plugins";
 import type Database from "metabase-lib/v1/metadata/Database";
-import type { Group, GroupsPermissions } from "metabase-types/api";
+import type {
+  DatabaseEntityId,
+  Group,
+  GroupsPermissions,
+  SpecialGroupType,
+} from "metabase-types/api";
 
 import { DATA_PERMISSION_OPTIONS } from "../../constants/data-permissions";
 import { Messages } from "../../constants/messages";
@@ -17,12 +22,7 @@ import {
   limitDatabasePermission,
   navigateToGranularPermissions,
 } from "../../permissions";
-import type {
-  DataPermissionValue,
-  DatabaseEntityId,
-  PermissionSectionConfig,
-  SpecialGroupType,
-} from "../../types";
+import type { DataPermissionValue, PermissionSectionConfig } from "../../types";
 import { DataPermission, DataPermissionType } from "../../types";
 import {
   getPermissionWarning,
