+++ frontend/src/metabase/admin/permissions/utils/data-entity-id.ts
@@ -2,14 +2,13 @@ import { checkNotNull } from "metabase/utils/types";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type Schema from "metabase-lib/v1/metadata/Schema";
 import type Table from "metabase-lib/v1/metadata/Table";
-import type { ConcreteTableId } from "metabase-types/api";
-
 import type {
+  ConcreteTableId,
   DatabaseEntityId,
-  EntityId,
+  PermissionEntityId,
   SchemaEntityId,
   TableEntityId,
-} from "../types";
+} from "metabase-types/api";
 
 export const getDatabaseEntityId = (databaseEntity: Database) => ({
   databaseId: databaseEntity.id,
@@ -27,11 +26,11 @@ export const getTableEntityId = (tableEntity: Table) => ({
 });
 
 export const isTableEntityId = (
-  entityId: Partial<EntityId>,
+  entityId: Partial<PermissionEntityId>,
 ): entityId is TableEntityId => entityId.tableId != null;
 
 export const isSchemaEntityId = (
-  entityId: Partial<EntityId>,
+  entityId: Partial<PermissionEntityId>,
 ): entityId is SchemaEntityId & { schemaName: string } =>
   // not sure why schemaName can be undefined on SchemaEntityId
   entityId.schemaName != null &&
@@ -39,7 +38,7 @@ export const isSchemaEntityId = (
   !isTableEntityId(entityId);
 
 export const isDatabaseEntityId = (
-  entityId: Partial<EntityId>,
+  entityId: Partial<PermissionEntityId>,
 ): entityId is DatabaseEntityId =>
   entityId.databaseId != null &&
   !isSchemaEntityId(entityId) &&
