+++ frontend/src/metabase/admin/permissions/utils/graph/data-permissions/get.ts
@@ -1,16 +1,14 @@
 import { getIn } from "icepick";
 
-import type {
-  DatabaseEntityId,
-  EntityId,
-  SchemaEntityId,
-  TableEntityId,
-} from "metabase/admin/permissions/types";
 import { PLUGIN_ADVANCED_PERMISSIONS } from "metabase/plugins";
 import {
   DataPermission,
   DataPermissionValue,
+  type DatabaseEntityId,
   type GroupsPermissions,
+  type PermissionEntityId,
+  type SchemaEntityId,
+  type TableEntityId,
 } from "metabase-types/api";
 
 // permission that do not have a nested schemas/native key
@@ -60,7 +58,7 @@ function getOmittedPermissionValue(
 export function getRawPermissionsGraphValue(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   permission: DataPermission,
 ) {
   const nestedPath = [
@@ -176,7 +174,7 @@ export const getFieldsPermission = (
 export const getEntityPermission = (
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   permission: DataPermission,
 ): DataPermissionValue => {
   if (entityId.tableId !== undefined) {
