+++ frontend/src/metabase/admin/permissions/utils/graph/data-permissions/update.ts
@@ -1,12 +1,6 @@
 import { getIn, setIn } from "icepick";
 import _ from "underscore";
 
-import type {
-  DatabaseEntityId,
-  EntityId,
-  SchemaEntityId,
-  TableEntityId,
-} from "metabase/admin/permissions/types";
 import {
   isSchemaEntityId,
   isTableEntityId,
@@ -21,7 +15,11 @@ import type Table from "metabase-lib/v1/metadata/Table";
 import {
   DataPermission,
   DataPermissionValue,
+  type DatabaseEntityId,
   type GroupsPermissions,
+  type PermissionEntityId,
+  type SchemaEntityId,
+  type TableEntityId,
 } from "metabase-types/api";
 
 import {
@@ -162,7 +160,7 @@ export function updateSchemasPermission(
 export function updateEntityPermission(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   value: DataPermissionValue,
   database: Database,
   permission: DataPermission,
@@ -200,7 +198,7 @@ export function updateEntityPermission(
 export function restrictCreateQueriesPermissionsIfNeeded(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   permission: DataPermission,
   value: DataPermissionValue,
   database: Database,
@@ -280,7 +278,7 @@ const hasFullCreateQueriesAccess = (
 export function revokeTransformsPermissionIfNeeded(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   permission: DataPermission,
   value: DataPermissionValue,
 ): GroupsPermissions {
@@ -334,7 +332,7 @@ export function revokeTransformsPermissionIfNeeded(
 export function revokeWorkspacesPermissionIfNeeded(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   permission: DataPermission,
   value: DataPermissionValue,
 ): GroupsPermissions {
@@ -388,7 +386,7 @@ export function revokeWorkspacesPermissionIfNeeded(
 function inferEntityPermissionValueFromChildTables(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   database: Database,
   permission: DataPermission,
 ): DataPermissionValue {
@@ -417,7 +415,7 @@ function inferEntityPermissionValueFromChildTables(
 export function inferAndUpdateEntityPermissions(
   permissions: GroupsPermissions,
   groupId: number,
-  entityId: EntityId,
+  entityId: PermissionEntityId,
   database: Database,
   permission: DataPermission,
 ) {
